requirements.txt
# Add your project dependencies here, one per line.
firecrawl
pydantic
click==8.1.7
sqlalchemy
psycopg2-binary
python-dotenv
openai
pinecone
beautifulsoup4
markdownify
pytest==8.4.2
pytest-cov==7.0.0
pytest-dotenv
scrapingbee
matplotlib
jinja2
onesignal-python-api>=5.3.0b1
cryptography
parsel
jq